pyproject.toml
[project]
name = "agentnn"
version = "1.0.0-beta"
description = "Modular agent execution and neural orchestration framework"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
  "fastapi",
  "httpx",
  "typer",
  "pydantic-settings",
  "pydantic",
  "PyYAML",
  "tqdm",
  "uvicorn",
  "slowapi",
  "aiofiles",
  "prometheus-client",
  "loguru",
  "python-json-logger",
  "aiohttp",
  "structlog",
  "mlflow"
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]

[project.scripts]
agentnn = "sdk.cli.main:app"
mcp_cli = "mcp.cli:app"

[tool.ruff]
line-length = 120
extend-exclude = [
  "archive",
  "agents",
  "api",
  "cli",
  "openhands_api",
  "llm_models",
  "nn_models",
  "managers",
  "plugins",
  "training",
  "datastores",
  "monitoring",
  "evaluation",
  "security",
  "rag",
  "mlflow_integration",
  "tests",
  "scripts",
  "benchmarks",
  "frontend",
  "docs",
]

[project.optional-dependencies]
sdk = []

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
